update movie api

+++ 05_Backend/API/movieAPI.js
@@ -3,8 +3,8 @@ const router = express.Router()
 const jwt = require("jsonwebtoken")
 const Movies = require("../schema/movies.js")
 const Users = require("../schema/users.js")
-const tokenValidator = require("../validator/tokenValidator.js")
-const checkRole = require("../validator/checkRole.js")
+const { check_admin, check_superAdmin } = require('../validator/checkRole')
+const { token_provided, verifyToken } = require('../validator/tokenValidator')
 
 // Get all movies
 router.get("/", async (req, res) => {
@@ -18,52 +18,78 @@ router.get("/", async (req, res) => {
     res.status(500).json({ error: "Internal server error" })
   }
 })
-// Insert a new movie
-router.post("/:id", async (req, res) => {
+
+// Get a specific movie by ID
+router.get("/:id", async (req, res) => {
   try {
-    // const id = req.params.id
-    // const token = req.headers.authorization
-    // // Check if the JWT token is provided
-    // if (!token) {
-    //   return res.status(401).json({ error: "Access denied. Token not provided." })
-    // }
-    // // Find the user by ID
-    // const userExist = await Users.findOne({ _id: id })
-    // const decodedToken = jwt.verify(
-    //   token.split(" ")[1],
-    //   process.env.SECRET_KEY
-    // )
-    // // Check if the user exists
-    // if (!userExist) {
-    //   return res.status(404).json({ error: "User not found." })
-    // }
-    // // Check if the user is authorized to create a movie
-    // if (
-    //   !(decodedToken.username === userExist.username) ||
-    //   !decodedToken.role === "superAdmin"
-    // ) {
-    //   return res.status(401).json({ error: "Sorry, you're not authorized." })
-    // }
+    const movieId = req.params.id
+    // Fetch the movie from the database by its ID
+    const movie = await Movies.findById(movieId)
+    // Check if the movie exists
+    if (!movie) {
+      return res.status(404).json({ error: "Movie not found" })
+    }
+    // Return the movie details in JSON format
+    return res.status(200).json(movie)
+  } catch (error) {
+    // Handle any errors that occur during fetching the movie
+    console.error(error)
+    return res.status(500).json({ error: "Internal server error" })
+  }
+})
 
-    //checking valid token
-    tokenValidator.verifyToken(req.headers.authorization)
+// Find a movie by name using query parameters
+// API Call 
+// http://localhost:5000/api/movies/search/name?name=Interstellar
+router.get("/search/name", async (req, res) => {
 
-    //check Super admin
-    if (checkRole.check_superAdmin(req.headers.authorization)) {
+  try {
+    const name = req.query.name;
 
-      console.log("Super Admin is found");
+    console.log("query " + name);
 
+    // Check if the name parameter is provided
+    if (!name) {
+      return res.status(400).send("Name parameter is missing")
     }
 
-    //check admin
-    if (checkRole.check_admin(req.headers.authorization)) {
+    // Find the movie by its name
+    const movie = await Movies.findOne({ name: name })
 
-      console.log("admin is found");
+    // Check if the movie exists
+    if (!movie) {
+      return res.status(404).send("Movie not found")
+    }
 
+    // Return the found movie
+    return res.status(200).json(movie);
+  } catch (error) {
+    console.log(error);
+    return res.status(500).send("Failed to find movie.")
+  }
+});
+
+
+
+// Insert a new movie
+router.post("/post", async (req, res) => {
+  try {
+
+    const token = req.headers.authorization
+    // Check if the JWT token is provided
+    if (!token_provided(token)) {
+      return res.status(401).json({ error: "Access denied. Token not provided." })
+    }
+    // Find the user by ID
+    const decodedToken = verifyToken(token)
+
+    // Check if the user is authorized to create a movie
+    if (!decodedToken || (!check_admin(token) && !check_superAdmin(token))) {
+      return res.status(401).json({ error: "Sorry, you're not authorized." })
     }
 
     // Create a new movie based on the request body
-    const newMovie = req.body
+    const newMovie = req.body;
     const createdMovie = await Movies.create(newMovie)
     // Return success message and the created movie
     return res.status(200).json({
@@ -75,47 +101,83 @@ router.post("/:id", async (req, res) => {
     return res.status(500).json({ error: "Failed to create movie." })
   }
 })
-// Update a movie by ID
-router.put("/:id", async (req, res) => {
+
+router.put("/update/:id", async (req, res) => {
   try {
     const id = req.params.id
     const updatedMovie = req.body
+    const token = req.headers.authorization
+
+    // Check if the JWT token is provided
+    if (!token_provided(token)) {
+      return res.status(401).json({ error: "Access denied. Token not provided." })
+    }
+
+    // Verify the token and extract the decoded token
+    const decodedToken = verifyToken(token)
+
+    // Check if the user is authorized to update a movie
+    if (!decodedToken || (!check_admin(token) && !check_superAdmin(token))) {
+      return res.status(401).json({ error: "Sorry, you're not authorized." })
+    }
+
     // Find and update the movie by its ID
     const movie = await Movies.findByIdAndUpdate(id, updatedMovie, {
-      new: true,
+      new: true
     })
+
     // Check if the movie exists
     if (!movie) {
       return res.status(404).json({ error: "Movie not found" })
     }
+
     // Return success message and the updated movie
     return res.status(200).json({
       message: "Movie updated successfully",
       movie: movie,
-    })
+    });
   } catch (error) {
     console.error(error)
     return res.status(500).json({ error: "Failed to update movie." })
   }
 })
+
 // Delete a movie by ID
-router.delete("/:id", async (req, res) => {
+router.delete("/delete/:id", async (req, res) => {
   try {
     const id = req.params.id
+    const token = req.headers.authorization
+
+    // Check if the JWT token is provided
+    if (!token_provided(token)) {
+      return res.status(401).json({ error: "Access denied. Token not provided." })
+    }
+
+    // Verify the token and extract the decoded token
+    const decodedToken = verifyToken(token)
+
+    // Check if the user is authorized to delete a movie
+    if (!decodedToken || (!check_admin(token) && !check_superAdmin(token))) {
+      return res.status(401).json({ error: "Sorry, you're not authorized." })
+    }
+
     // Find and delete the movie by its ID
     const deletedMovie = await Movies.findByIdAndDelete(id)
+
     // Check if the movie exists
     if (!deletedMovie) {
       return res.status(404).json({ error: "Movie not found" })
     }
+
     // Return success message and the deleted movie
     return res.status(200).json({
       message: "Movie deleted successfully",
       movie: deletedMovie,
-    })
+    });
   } catch (error) {
     console.error(error)
     return res.status(500).json({ error: "Failed to delete movie." })
   }
 })
+
 module.exports = router
